DEV: Use a scheduled job to update stats in bg
This is a follow up to previous commit to ensure that metric collection
is not slow for multisites with hundreds of sites.
