+++ spec/lib/internal_metric/global_spec.rb
@@ -104,14 +104,12 @@ RSpec.describe DiscoursePrometheus::InternalMetric::Global do
     end
   end
 
-  it "collects pg_seq metric and caches the sequence" do
+  it "collects postgres_highest_sequence metric" do
+    Jobs::UpdateStats.new.execute
+
     metric.collect
 
     expect(metric.postgres_highest_sequence).to be_a_kind_of(Hash)
     expect(metric.postgres_highest_sequence[{ db: "default" }]).to be_present
-
-    expect do metric.collect end.not_to change {
-      metric.class.class_variable_get(:@@postgres_highest_sequence_last_check)
-    }
   end
 end
